Migrate to use directive resolvers instead of shield middleware
