+++ graphql/package.json
@@ -34,7 +34,6 @@
     "esm": "^3.2.14",
     "graphql": "^14.0.2",
     "graphql-import": "^0.7.1",
-    "graphql-shield": "^5.3.4",
     "graphql-yoga": "^1.16.7",
     "graphqlgen": "^0.4.0",
     "jsonwebtoken": "^8.4.0",
