+++ graphql/src/helpers/headers/index.ts
@@ -3,3 +3,5 @@ export * from './getTokenFromRequest';
 export { default as setTokenToResponse } from './setTokenToResponse';
 export * from './setTokenToResponse';
 export { default as TokenType } from './TokenType';
+export { default as getViewerFromRequest } from './getViewerFromRequest';
+export * from './getViewerFromRequest';
