--- graphql/src/middleware/shield/index.ts
@@ -1,2 +1,0 @@
-export { default } from './shield';
-export * from './shield';
