+++ graphql/src/server.ts
@@ -5,17 +5,18 @@ import { GraphQLServer } from 'graphql-yoga';
 import { getTokenFromRequest, setTokenToResponse } from 'src/helpers/headers';
 import { sign, verify } from 'src/helpers/utils/jwt';
 import { compare, hash } from 'src/helpers/utils/password';
-import shield, { getViewerFromRequest } from 'src/middleware/shield';
 import tokenVerifier from 'src/middleware/tokenVerifier';
 import { prisma } from 'src/web/graphql/generated/prisma-client';
 import { resolvers } from 'src/web/graphql/resolvers';
+import directiveResolvers from 'src/web/graphql/directiveResolvers';
+import { getViewerFromRequest } from 'src/helpers/headers';
 
 const PORT = process.env.PORT || 8000;
 
 const server = new GraphQLServer({
   typeDefs: importSchema('./src/schemas/schema.graphql'),
   resolvers: resolvers as any, // https://github.com/prisma/graphql-yoga/issues/379
-  middlewares: [shield],
+  directiveResolvers,
   context: async contextParams => ({
     ...contextParams,
     prisma,
@@ -49,7 +50,5 @@ server.start(
     formatError,
     cors: { origin: process.env.SERVER_NAME, credentials: true }
   },
-  () =>
-    // tslint:disable-next-line
-    console.log(`Server is running on http://localhost:${PORT}`)
+  () => console.log(`Server is running on http://localhost:${PORT}`)
 );
