+++ graphql/src/web/graphql/directiveResolvers/index.ts
@@ -1,0 +1,6 @@
+import isOwner from './isOwner';
+import isLoggedIn from './isLoggedIn';
+
+const directiveResolvers = { isOwner, isLoggedIn };
+
+export default directiveResolvers;
